+++ .idea/misc.xml
@@ -38,6 +38,10 @@
         <entry key="../../../../../../layout/compose-model-1653206065009.xml" value="0.18703703703703703" />
         <entry key="../../../../../../layout/compose-model-1653473003135.xml" value="0.1638888888888889" />
         <entry key="../../../../../../layout/compose-model-1653473093021.xml" value="0.375" />
+        <entry key="../../../../../../layout/compose-model-1653473952484.xml" value="0.1638888888888889" />
+        <entry key="../../../../../../layout/compose-model-1653473972839.xml" value="1.0" />
+        <entry key="../../../../../../layout/compose-model-1653474106565.xml" value="0.1638888888888889" />
+        <entry key="../../../../../../layout/compose-model-1653485614547.xml" value="0.1" />
         <entry key="common/src/main/res/drawable/ic_launcher_background.xml" value="0.1185" />
       </map>
     </option>
