+++ common/build.gradle.kts
@@ -41,6 +41,17 @@ dependencies {
     androidTestImplementation(Test.JUNIT_EXT)
     androidTestImplementation(Test.ESPRESSO)
 
+    // Compose
+    implementation(Compose.COMPOSE_ACTIVITY)
+    implementation(Compose.COMPOSE_UI)
+    implementation(Compose.COMPOSE_MATERIAL)
+    implementation(Compose.COMPOSE_TOOLING_PREVIEW)
+    implementation(Compose.COMPOSE_LIFECYCLE)
+    implementation(Compose.COMPOSE_NAVIGATION)
+    implementation(Compose.COMPOSE_HILT)
+    debugImplementation(Compose.COMPOSE_TOOLING)
+    debugImplementation(Compose.COMPOSE_TEST_MANIFEST)
+
     // Timber logging
     implementation(Log.TIMBER)
 }
