+++ common/src/main/java/com/andriawan/common/navigation/NavigationExt.kt
@@ -1,4 +1,4 @@
-package com.andriawan.template.utils
+package com.andriawan.common.navigation
 
 import androidx.navigation.NavHostController
