+++ common/src/main/java/com/andriawan/common/navigation/Routes.kt
@@ -1,4 +1,4 @@
-package com.andriawan.common
+package com.andriawan.common.navigation
 
 object Routes {
     const val HOME_PAGE = "home"
